requirements.txt
# Weather Collector Dependencies
#
# Copyright (c) 2025 Darren Soothill
# Email: darren [at] soothill [dot] com
# All rights reserved.

# HTTP client for API requests
requests==2.32.5

# InfluxDB v2 Python client
influxdb-client==1.49.0

# YAML configuration file parsing
PyYAML==6.0.3

# Date and time handling
python-dateutil==2.9.0.post0

# System monitoring for performance testing
psutil==6.1.1

# Threading utilities (built-in, but listed for clarity)
# threading is part of standard library
